move shared files and directories back to collection-items-panel
