+++ src/sections/account/my-data-section/MyDataItemsPanel.tsx
@@ -3,12 +3,12 @@ import { Stack } from '@iqss/dataverse-design-system'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
 import { CollectionItemsPaginationInfo } from '@/collection/domain/models/CollectionItemsPaginationInfo'
 import { CollectionItemType } from '@/collection/domain/models/CollectionItemType'
-import { useLoadMoreOnPopStateEvent } from '../../shared/collection-items-panel/useLoadMoreOnPopStateEvent'
+import { useLoadMoreOnPopStateEvent } from '../../collection/collection-items-panel/useLoadMoreOnPopStateEvent'
 import { useLoading } from '@/sections/loading/LoadingContext'
-import { ItemsList } from '../../shared/collection-items-panel/items-list/ItemsList'
+import { ItemsList } from '@/sections/collection/collection-items-panel/items-list/ItemsList'
 import { MyDataFilterPanel } from '@/sections/account/my-data-section/my-data-filter-panel/MyDataFilterPanel'
-import { SearchPanel } from '../../shared/collection-items-panel/search-panel/SearchPanel'
-import { ItemTypeChange } from '../../shared/collection-items-panel/filter-panel/type-filters/TypeFilters'
+import { SearchPanel } from '@/sections/collection/collection-items-panel/search-panel/SearchPanel'
+import { ItemTypeChange } from '@/sections/collection/collection-items-panel/filter-panel/type-filters/TypeFilters'
 import { MyDataSearchCriteria } from '@/sections/account/my-data-section/MyDataSearchCriteria'
 import { useGetMyDataAccumulatedItems } from '@/sections/account/my-data-section/useGetMyDataAccumulatedItems'
 import { RoleChange } from '@/sections/account/my-data-section/my-data-filter-panel/role-filters/RoleFilters'
