+++ src/sections/account/my-data-section/my-data-filter-panel/MyDataFilterPanel.tsx
@@ -4,10 +4,10 @@ import { Button, Offcanvas } from '@iqss/dataverse-design-system'
 import { FunnelFill } from 'react-bootstrap-icons'
 import { CollectionItemType } from '@/collection/domain/models/CollectionItemType'
 import { Role } from '@/users/domain/models/Role'
-import { ItemTypeChange } from '@/sections/shared/collection-items-panel/filter-panel/type-filters/TypeFilters'
+import { ItemTypeChange } from '@/sections/collection/collection-items-panel/filter-panel/type-filters/TypeFilters'
 import { PublicationStatusChange } from '@/sections/account/my-data-section/my-data-filter-panel/publication-status-filters/PublicationStatusFilters'
 import { RoleChange } from '@/sections/account/my-data-section/my-data-filter-panel/role-filters/RoleFilters'
-import { TypeFilters } from '@/sections/shared/collection-items-panel/filter-panel/type-filters/TypeFilters'
+import { TypeFilters } from '@/sections/collection/collection-items-panel/filter-panel/type-filters/TypeFilters'
 import { RoleFilters } from '@/sections/account/my-data-section/my-data-filter-panel/role-filters/RoleFilters'
 
 import styles from './MyDataFilterPanel.module.scss'
