+++ src/sections/collection/collection-items-panel/CollectionItemsPanel.tsx
@@ -11,17 +11,17 @@ import {
 } from '@/collection/domain/models/CollectionSearchCriteria'
 import { CollectionItemType } from '@/collection/domain/models/CollectionItemType'
 import { CollectionItemsQueryParams } from '@/collection/domain/models/CollectionItemsQueryParams'
-import { useGetAccumulatedItems } from '../../shared/collection-items-panel/useGetAccumulatedItems'
+import { useGetAccumulatedItems } from './useGetAccumulatedItems'
 import { UseCollectionQueryParamsReturnType } from '../useGetCollectionQueryParams'
-import { useLoadMoreOnPopStateEvent } from '../../shared/collection-items-panel/useLoadMoreOnPopStateEvent'
+import { useLoadMoreOnPopStateEvent } from './useLoadMoreOnPopStateEvent'
 import { useLoading } from '@/sections/loading/LoadingContext'
 import { CollectionHelper } from '../CollectionHelper'
-import { FilterPanel } from '../../shared/collection-items-panel/filter-panel/FilterPanel'
-import { RemoveAddFacetFilter } from '@/sections/shared/collection-items-panel/filter-panel/facets-filters/FacetFilterGroup'
-import { ItemsList } from '../../shared/collection-items-panel/items-list/ItemsList'
-import { SearchPanel } from '../../shared/collection-items-panel/search-panel/SearchPanel'
-import { ItemTypeChange } from '../../shared/collection-items-panel/filter-panel/type-filters/TypeFilters'
-import { SelectedFacets } from '../../shared/collection-items-panel/selected-facets/SelectedFacets'
+import { FilterPanel } from '@/sections/collection/collection-items-panel/filter-panel/FilterPanel'
+import { RemoveAddFacetFilter } from '@/sections/collection/collection-items-panel/filter-panel/facets-filters/FacetFilterGroup'
+import { ItemsList } from '@/sections/collection/collection-items-panel/items-list/ItemsList'
+import { SearchPanel } from '@/sections/collection/collection-items-panel/search-panel/SearchPanel'
+import { ItemTypeChange } from '@/sections/collection/collection-items-panel/filter-panel/type-filters/TypeFilters'
+import { SelectedFacets } from '@/sections/collection/collection-items-panel/selected-facets/SelectedFacets'
 import styles from './CollectionItemsPanel.module.scss'
 
 interface CollectionItemsPanelProps {
