+++ src/sections/collection/collection-items-panel/filter-panel/FilterPanel.module.scss
