+++ src/sections/collection/collection-items-panel/filter-panel/facets-filters/FacetFilterGroup.tsx
