+++ src/sections/collection/collection-items-panel/filter-panel/facets-filters/FacetsFilters.module.scss
