+++ src/sections/collection/collection-items-panel/filter-panel/type-filters/TypeFilters.tsx
