+++ src/sections/collection/collection-items-panel/items-list/ItemsList.module.scss
