+++ src/sections/collection/collection-items-panel/items-list/collection-card/CollectionCard.module.scss
