+++ src/sections/collection/collection-items-panel/items-list/file-card/FileCard.module.scss
