+++ src/sections/collection/collection-items-panel/items-list/file-card/FileCardThumbnail.tsx
