+++ src/sections/collection/collection-items-panel/search-panel/SearchPanel.module.scss
