+++ src/sections/collection/collection-items-panel/selected-facets/SelectedFacets.module.scss
