+++ src/sections/collection/collection-items-panel/selected-facets/SelectedFacets.tsx
