+++ src/sections/collection/collection-items-panel/useLoadMoreOnPopStateEvent.ts
