+++ tests/component/sections/collection/collection-items-panel/ItemsSortBy.spec.tsx
@@ -1,4 +1,4 @@
-import { ItemsSortBy } from '@/sections/shared/collection-items-panel/items-list/ItemsSortBy'
+import { ItemsSortBy } from '@/sections/collection/collection-items-panel/items-list/ItemsSortBy'
 import { OrderType, SortType } from '@/collection/domain/models/CollectionSearchCriteria'
 
 describe('ItemsSortBy', () => {
